updating minor naming conventions for vars and types on tests

+++ src/blockchain/smilobft/consensus/sport/smilobftcore/core_handler_test.go
@@ -27,7 +27,7 @@ import (
 
 // notice: the normal case have been tested in integration tests.
 func TestHandleMsg(t *testing.T) {
-	N := uint64(4)
+	N := 4
 	sys := NewTestSystemWithBackend(N)
 
 	closer := sys.Run(true)
